==> Analysis/2_Mouse-data-preprocessing-PCIbex.Rmd <==
---
title: "Mouse-data-preprocessing-PCIbex"
output: html_document
date: "2022-08-09"
author: "Swarnendu Moitra, Indranil Dutta"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(dplyr)
```


## Instructions from the website

newMouseTracker("mouse")

The only parameter is the element name ("mouse") which is optional. Make sure to call log to collect data in the results file.

**Note 1:** The first four numbers (preceded by w, h, x and y) report the width and height of the page, and the x/y coordinates of the mouse when the MouseTracker element’s start command is executed. After that, the string is a series of triplets of numbers, each triplet is prefixed with t and the numbers are separated by +/-: the first number is the number of milliseconds that have elapsed since the previous triplet (t is for “time”), the second number is how many pixels the mouse moved horizontally (+ = right, - = left) and the third number is how many pixels the mouse moved vertically (+ = down, - = up)

**Note 2:** The coordinates are reported in the results file as a long compressed string. The string always start with **xXyYwWhH**, where X and Y are numbers corresponding to the mouse coordinates when recording starts, and W and H correspond to the page’s width and height. It then continues with sequences of tT+A-B bits, where T corresponds to the number of milliseconds since the last bit, and +A and -B, which can be positive (+) or negative (-) correspond to how many pixels the mouse has moved horizontally and vertically, respectively. This script illustrates how to decompress the string in R, assuming the results datatable was previously subset to MouseTracker rows only:

https://www.pcibex.net/wiki/mousetracker-element/

```{r}
library(tidyverse)
```

# Preprocessing Starts here.

## Subset the parsed Results/Data CSV containing only the mouse-tracking data rows

```{r}
rawfile <- read.csv("data/Data.csv")

#results2 <- filter(file, PennElementType == "MouseTracker")
results <- filter(rawfile, Parameter == "Move")
```


```{r}
timestamps_get_response <- vector()
xpos_get_response <- vector()
ypos_get_response <- vector()
for(row in 1:nrow(results)){
  time <- 0
  stream <- as.character(results[row,"Value"])
  pos <- data.frame(time=c(time),x=as.numeric(gsub("^x(\\d+)y.+$","\\1", stream)), y=as.numeric(gsub("^.+y(\\d+)w.+$","\\1", stream)))
  ptime <- time
  counter <- 1
  px <- pos[1,'x']
  py <- pos[1,'y']
  for(s in (strsplit(stream,'t')[[1]][-1])){
    row <- strsplit(gsub("^(\\d+)([+-]\\d+)([+-]\\d+)$","\\1 \\2 \\3",s),' ')
    ntime <- as.numeric(ptime+as.numeric(row[[1]][1]))
    nx <- as.numeric(px+as.numeric(row[[1]][2]))
    ny <- as.numeric(py+as.numeric(row[[1]][3]))
    pos <- rbind(pos, data.frame(time=ntime,x=nx,y=ny))
    ptime <- ntime
    px <- nx
    py <- ny
    counter <- counter +1
  }
  times <- toString(pos[["time"]])
  timestamp <- paste("[", times, "]", sep = "")
  timestamps_get_response <- c(timestamps_get_response, timestamp)
  xpos <- toString(pos[["x"]])
  xpos_str <- paste("[", xpos, "]", sep = "")
  xpos_get_response <- c(xpos_get_response, xpos_str)
  ypos <- toString(pos[["y"]])
  ypos_str <- paste("[", ypos, "]", sep = "")
  ypos_get_response <- c(ypos_get_response, ypos_str)
}
results$timestamps <- timestamps_get_response
results$xpos <- xpos_get_response
results$ypos <- ypos_get_response
```

```{r}
# inner_join based on X
library(dplyr)
results$xpos <- str_sub(results$xpos, 2, -2)
results$ypos <- str_sub(results$xpos, 2, -2)
results$Timestamps <- str_sub(results$timestamps, 2, -2)
results$xpos <-str_replace_all(results$xpos, " ", "")
results$ypos <-str_replace_all(results$xpos, " ", "")
results$Timestamps <-str_replace_all(results$Timestamps, " ", "")

#results <- results %>% select(X,xpos,ypos,Timestamps)
#output <- merge (x = rawfile, y = results, by = "X", all.x = TRUE)
```


```{r}
#xs$xpos <- str_sub(xs$xpos, 2, -2)

#xs$all <- str_replace_all(xs$xpos, " ", "")
```

```{r}
#write.csv(output,"data/MPhil-MT-Data.csv")
#write.csv2(output, "data/MPhil-MT_Data2.csv")

write.csv(results,"data/MPhil-MT-Data.csv")
write.csv2(results, "data/MPhil-MT_Data2.csv")
```
